# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: discogs-fade-backend
  version: 0.1.0
  description: >
    Boundary between the extension and Discogs' API. Wraps GET /masters/{id} and
    GET and POST /fade — master DTOs are currently unscoped (all Discogs response
    fields kept as-is). GET /releases/{id} serves a trimmed release from our
    own database, and GET /releases lists and filters the releases stored there.
paths:
...
  /fade:
    get:
      operationId: getFades
      summary: List what the signed-in user has faded
      description: >
        Faded masters and releases without a master. A release is hidden when
        its own id is in releaseIds or its master id is in masterIds, which
        also covers versions added to a faded master later. versionReleaseIds
        lists the versions of the faded masters found so far, and
        failedLookups the fades Discogs could not resolve.
      responses:
        "200":
          description: Faded ids
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FadedIdsDto"
    post:
      operationId: fade
      summary: >
        Fade a release or a master for the signed-in user. Any release or
        master can be faded, indexed or not. A release with a known master is
        stored as that master, hiding every version under it. Any other
        release is stored as itself, and a lookup (queued ahead of scans and
        enrichment) upgrades it to its master and stores the master's
        versions. The request waits for that lookup (up to about 20 seconds),
        so the answer names the fade and counts what it hides. If the wait
        runs out, or Discogs is paused, lookupStatus is pending, versionCount
        is null and the lookup carries on in the background. Fading a failed
        fade again retries its lookup. Otherwise idempotent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FadeRequestDto"
      responses:
        "200":
          description: Fade stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FadeResponseDto"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    delete:
      operationId: unfade
      summary: >
        Unfade a release or a master for the signed-in user. A release with a
        master unfades that master and every version under it. Releases that
        the fade kept out of enrichment are queued for enrichment again, unless
        another fade of the user still covers them. Idempotent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FadeRequestDto"
      responses:
        "204":
          description: Fade removed, or there was none
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /fade/items:
    get:
      operationId: getFadedItems
      summary: Page through what the signed-in user has faded
      description: >
        One row per stored fade (a master, or a release without a master),
        newest first, each with the lowest-id indexed release it hides.
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: >
            Case-insensitive substring of a title or artist name of any indexed
            release the fade hides.
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, default: 50 }
      responses:
        "200":
          description: A page of fades
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FadedItemsDto"
        "400":
          description: Invalid page or pageSize
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
...
components:
...
  schemas:
    ErrorDto:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    FadeRequestDto:
      type: object
      description: Provide exactly one of releaseId or masterId.
      properties:
        releaseId:
          type: integer
        masterId:
          type: integer
    FadeResponseDto:
      type: object
      required: [kind, id, lookupStatus, title, artists, versionCount]
      properties:
        kind:
          type: string
          enum: [master, release]
        id:
          type: integer
        lookupStatus:
          type: string
          enum: [pending, done, failed]
          description: >
            pending until the master and its versions are known, done once they
            are (or the release has no master), failed when Discogs could not
            resolve it.
        title:
          type: string
          nullable: true
          description: The title of the faded release or master, when known.
        artists:
          type: array
          items: { $ref: "#/components/schemas/ReleaseArtistStub" }
        versionCount:
          type: integer
          nullable: true
          description: >
            Releases the fade hides: the master's versions, or 1 for a release
            without a master. Null unless lookupStatus is done.
...
    FadedItemDto:
      type: object
      required: [kind, id, fadedAt, title, artists, year, thumb, versionsIndexed, lookupStatus]
      properties:
        kind:
          type: string
          enum: [master, release]
        id:
          type: integer
          description: The master id or release id, as stored; send it back to unfade.
        fadedAt:
          type: string
          format: date-time
        title:
          type: string
          nullable: true
          description: >
            Of the lowest-id indexed release the fade hides, else the title
            Discogs gave when the fade was looked up; null if neither is known.
        artists:
          type: array
          items: { $ref: "#/components/schemas/ReleaseArtistStub" }
        year:
          type: integer
          nullable: true
        thumb:
          type: string
          nullable: true
        versionsIndexed:
          type: integer
          description: Indexed releases the fade hides; 0 when none is indexed.
        lookupStatus:
          type: string
          enum: [pending, done, failed]
          description: Where the fade's Discogs lookup stands; failed when Discogs could not resolve it.
    FadedItemsDto:
      type: object
      required: [items, page, pageSize, total, fadedTotal]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/FadedItemDto" }
        page: { type: integer }
        pageSize: { type: integer }
        total:
          type: integer
          description: Fades matching the search.
        fadedTotal:
          type: integer
          description: All of the collector's fades, ignoring the search.
    FadedIdsDto:
      type: object
      required: [masterIds, releaseIds, versionReleaseIds, failedLookups]
      properties:
        masterIds:
          type: array
          items: { type: integer }
        releaseIds:
          type: array
          items: { type: integer }
        versionReleaseIds:
          type: array
          items: { type: integer }
          description: Release ids under the faded masters, as far as their lookups have run.
        failedLookups:
          type: array
          description: Fades Discogs could not resolve. They stay in place; fading again retries.
          items:
            type: object
            required: [kind, id]
            properties:
              kind:
                type: string
                enum: [master, release]
              id:
                type: integer
...
    ReleaseArtistStub:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
...